refactor(edge/update): break utils to small files
remove utils
